[COO] Add rotation to Carre & Losange

+++ First Half/COO/TP-2/src/multifonctionnelles/Carre.java
@@ -12,8 +12,8 @@ public class Carre extends Forme {
 	
 	@Override
 	public void dessiner () {
-		int x = xCentre - (taille / 2);
-		int y = yCentre + (taille / 2);
+		double x = xCentre - (taille / 2);
+		double y = yCentre + (taille / 2);
 		machineTrace.placer(rotationX(x, y), rotationY(x, y));
 		machineTrace.baisser();
 		y -= taille;
